Finish replacing of legacy lifecycle methods
