+++ components/at_mention/at_mention.jsx
@@ -54,25 +54,17 @@ export default class AtMention extends React.PureComponent {
         this.state = {
             user: getUserFromMentionName(props),
             show: false,
-            mentionName: '',
             usersByUsername: '',
         };
 
         this.overlayRef = React.createRef();
     }
 
-    // UNSAFE_componentWillReceiveProps(nextProps) { // eslint-disable-line camelcase
-    //     if (nextProps.mentionName !== this.props.mentionName || nextProps.usersByUsername !== this.props.usersByUsername) {
-    //         this.setState({
-    //             user: this.getUserFromMentionName(nextProps),
-    //         });
-    //     }
-    // }
-
     static getDerivedStateFromProps(nextProps, prevState) {
-        if (nextProps.mentionName !== prevState.mentionName || nextProps.usersByUsername !== prevState.usersByUsername) {
+        if (nextProps.usersByUsername !== prevState.usersByUsername) {
             return {
                 user: getUserFromMentionName(nextProps),
+                usersByUsername: nextProps.usersByUsername,
             };
         }
 
@@ -89,26 +81,6 @@ export default class AtMention extends React.PureComponent {
     hideOverlay = () => {
         this.setState({show: false});
     }
-
-    // getUserFromMentionName(props) {
-    //     const usersByUsername = props.usersByUsername;
-    //     let mentionName = props.mentionName.toLowerCase();
-
-    //     while (mentionName.length > 0) {
-    //         if (usersByUsername.hasOwnProperty(mentionName)) {
-    //             return usersByUsername[mentionName];
-    //         }
-
-    //         // Repeatedly trim off trailing punctuation in case this is at the end of a sentence
-    //         if ((/[._-]$/).test(mentionName)) {
-    //             mentionName = mentionName.substring(0, mentionName.length - 1);
-    //         } else {
-    //             break;
-    //         }
-    //     }
-
-    //     return '';
-    // }
 
     render() {
         if (!this.state.user) {
