+++ components/edit_post_modal/edit_post_modal.jsx
@@ -60,14 +60,6 @@ export default class EditPostModal extends React.PureComponent {
         };
     }
 
-    // UNSAFE_componentWillUpdate(nextProps) { // eslint-disable-line camelcase
-    //     if (!this.props.editingPost.show && nextProps.editingPost.show) {
-    //         this.setState({
-    //             editText: nextProps.editingPost.post.message_source || nextProps.editingPost.post.message,
-    //         });
-    //     }
-    // }
-
     updateComponentWithEditText = (prevProps) => {
         if (!prevProps.editingPost.show && this.props.editingPost.show) {
             this.setState({
