+++ components/file_attachment/file_attachment.jsx
@@ -62,22 +62,13 @@ export default class FileAttachment extends React.PureComponent {
         this.loadFiles();
     }
 
-    // UNSAFE_componentWillReceiveProps(nextProps) { // eslint-disable-line camelcase
-    //     if (nextProps.fileInfo.id !== this.props.fileInfo.id) {
-    //         const extension = nextProps.fileInfo.extension;
-
-    //         this.setState({
-    //             loaded: getFileType(extension) !== FileTypes.IMAGE && !(nextProps.enableSVGs && extension === FileTypes.SVG),
-    //         });
-    //     }
-    // }
-
     static getDerivedStateFromProps(nextProps, prevState) {
         if (nextProps.fileInfo.id !== prevState.fileInfo.id) {
             const extension = nextProps.fileInfo.extension;
 
             return {
                 loaded: getFileType(extension) !== FileTypes.IMAGE && !(nextProps.enableSVGs && extension === FileTypes.SVG),
+                fileInfo: nextProps.fileInfo,
             };
         }
