+++ components/post_view/post_body/post_body.jsx
@@ -5,6 +5,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import {Posts} from 'mattermost-redux/constants';
 
+import LoadingSpinner from 'components/widgets/loading/loading_spinner';
 import * as PostUtils from 'utils/post_utils.jsx';
 import * as Utils from 'utils/utils.jsx';
 import DelayedAction from 'utils/delayed_action';
@@ -15,7 +16,6 @@ import FailedPostOptions from 'components/post_view/failed_post_options';
 import PostBodyAdditionalContent from 'components/post_view/post_body_additional_content';
 import PostMessageView from 'components/post_view/post_message_view';
 import ReactionList from 'components/post_view/reaction_list';
-import LoadingSpinner from 'components/widgets/loading/loading_spinner';
 
 const SENDING_ANIMATION_DELAY = 3000;
 
@@ -108,23 +108,12 @@ export default class PostBody extends React.PureComponent {
         this.sendingAction.cancel();
     }
 
-    // UNSAFE_componentWillReceiveProps(nextProps) { // eslint-disable-line camelcase
-    //     const post = nextProps.post;
-    //     if (post && post.id !== post.pending_post_id) {
-    //         this.sendingAction.cancel();
-    //         this.setState({sending: false});
-    //     }
-    // }
-
-    static getDerivedStateFromProps(nextProps) {
-        const post = nextProps.post;
+    componentDidUpdate() {
+        const post = this.props.post;
         if (post && post.id !== post.pending_post_id) {
-            // this.sendingAction.cancel();
-            return {
-                sending: false,
-            };
+            this.sendingAction.cancel();
+            this.setState({sending: true});
         }
-        return null;
     }
 
     render() {
